requirements.txt
groq==0.13.0
crewai==0.35.0
spacy==3.7.2
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
pydantic==2.7.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
